Fix fp16 NaN in multinomial sampling at long context
Cast logits to float32 before logits_processor to prevent NaN/Inf
from fp16 precision loss at 122K context causing multinomial failure.

+++ sd_code/hl/speculative/speculative_decoding.py
@@ -161,7 +161,7 @@ def speculative_generate(
             if use_greedy_sampler:
                 t = torch.argmax(last_logits.squeeze(1), dim=-1)
             else:
-                p_p = logits_processor(last_logits.squeeze(1))
+                p_p = logits_processor(last_logits.squeeze(1).float())
                 t = logits_processor.sample(p_p)
             del last_logits
         else:
@@ -283,7 +283,7 @@ def speculative_generate(
                     break
                 drafts_accepted += 1
         else:
-            p = logits_processor(target_logits[0])
+            p = logits_processor(target_logits[0].float())
             r = torch.rand(corrected_gamma, device=target.device)
             p = p.float()
             fractions = p[:corrected_gamma, ...] / q
